rf change tests dependable on ManageJenkins
