+++ src/test/java/school/redrover/ManageJenkinsTest.java
@@ -5,6 +5,7 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.testng.Assert;
 import org.testng.annotations.DataProvider;
+import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 import school.redrover.model.*;
 import school.redrover.runner.BaseTest;
@@ -25,15 +26,15 @@ public class ManageJenkinsTest extends BaseTest {
 
     @Test
     public void testSearchWithLetterConfigureSystem() {
-        String configurePage = new ManagePage(getDriver())
-                .navigateToManagePage()
-                .enterSearchQuery("m")
-                .clickSearchButton()
+        String configurePage = new MainPage(getDriver())
+                .navigateToManageJenkinsPage()
+                .inputToSearchField("m")
                 .selectOnTheFirstLineInDropdown()
                 .getConfigureSystemPage();
 
         Assert.assertEquals(configurePage, "Configure System");
     }
+
 
     @Test
     public void testManageConfigureNumberOfExecutorsInMasterNode() {
@@ -50,13 +51,12 @@ public class ManageJenkinsTest extends BaseTest {
         Assert.assertEquals(number, numberInLine);
     }
 
-    @Test
-    public void testBreadcrumbNavigateManageJenkins() {
 
-        String page = new ManagePage(getDriver())
-                .navigateToDashboardIcon()
-                .dropdownBreadcrumps()
-                .navigateToManageJenkinsAndClick()
+    @Test
+    public void testNavigateToManageJenkinsfromMainPageUsingDashboard() {
+
+        String page = new MainPage(getDriver())
+                .clickManageJenkinsOnDropDown()
                 .verifyManageJenkinsPage();
 
         Assert.assertEquals(page, "Manage Jenkins");
